Add behaviour to sanitise everything
This commit adds the start of the API to sanitise the database. The
`sanitise_table` method allows provides a method to sanitise individual
columns in the table. The columns can either be replaced with a string
(denoted by the `string` method), or be a dynamic bit of SQL that can
use functions and other data to generate a new value for the column.