Sanitise rows in table based on filter
Sometimes we want to sanitise rows in a table differently depending on
some condition. This commit adds a `where` option to the
`sanitise_table` declarations which allows an ActiveRecord-style query
to be added, which will limit the sanitisation to those rows.

This commit also adds a test to make sure that tables can be sanitised
multiple times to make sure it's possible to sanitise all rows in a
table, even when filtering is applied.