Raise an error if all columns aren't being sanitised
If a new column is added to the schema, it's easy to forget to check if
it needs to be sanitised. To prevent this, this commit raises an error
if not all columns are listed as being sanitised.

This will provide some friction to the development process, but it means
that developers will have to make a decision about what needs to happen
to a column when they create it.