Allow columns to be ignored when validating schema
The previous commit added an error if not all columns are accounted for
when sanitising values. This would mean that you would have to add
all columns and just return its value.

Instead, columns can be ignored, which means that they won't be
sanitised but they won't cause an error.