Extract sanitisation code into strategy
We want to be able to validate the entire configuration without running
the sanitisation. This means that we could add a really quick check to
the test suite to make sure that no columns or tables have been added
that haven't been accounted for.

This commit moves all of the code that actually executes the
sanitisation into a strategy. This will allow us to add other strategies
which could do other things like just validation.

At the moment, the column validation is still in the SanitiseDsl which
means there is some duplication of the code to create ActiveRecord
classes, but this may disappear once we have other strategies.