Add option to see what will happen during sanitisation
It would be useful to see what will happen if we run sanitisation
without actually running it. This would allow us to check coverage and
see the queries that will be executed.

This commit adds a `dry_run` method which will print the queries to any
IO interface.

I can't find a nice way of getting the SQL of an update statement in
Rails, so for the sanitisation it prints a select statement showing what
rows will be matched.