Ensure DbSanitiser has to be enabled before use
There will be some databases that should not be sanitised, for instance
a production database.

This commit adds a check to ensure that DbSanitiser is enabled before it
can be used. It is disabled by default so it can't be used accidentally.

In a Rails app, it could be enabled in an initializer that checks that
the environment isn't production before turning the DbSanitiser on.