Add partial deletion of table
There are some cases where deleting some rows of a table are useful.
This commit adds the functionality to make this happen.

The `partially_delete` method works a lot like sanitisation - it
requires all columns to be `allow`ed, so that any changes to the table
will require a developer to decide if the new column contains data that
needs to be sanitised in a different way.