Switch tests to use mysql
I'd like to introduce and test some optimisations to the sanitise
strategy that rely on behaviour specific to MySQL. In some cases
equivalents are available for SQLite, but having database-specific
strategies just for testing seems unjustifiable given that we only ever
use this with MySQL.

This therefore switches us over to test against MySQL instead. I've made
some assumptions about the DB configuration (namely that there's a
root-accessible passwordless DB sitting on localhost), largely in the
interests of speed and minimal configuration hassle. If someone does
ever want to test the gem with a different setup, we can look at
creating a customisable example config instead.