Add drop_and_create_index to speed large table updates
Some of our tables take ages to sanitise on staging; primarily, those
updating large tables that have string columns with indexes on them.

Testing suggests that these updates would be much quicker if we first
dropped the index, then performed the update, and finally recreated the
index.

This adds the relevant functionality to the DSL, and implements it in
both the sanitise and dry-run strategies.

I haven't done anything with the validate strategy yet, because I think
the best check is to verify that the schema doesn't change after the
drop-and-create. I'll implement this in the next commit.