Add separate truncate option to DB sanitiser
In a previous commit I switched the `delete_all` option to use a
TRUNCATE query for performance reasons. (cac669f)

However, it turns out this causes an error if used on a table with a
foreign key constraint pointed at it, even if the removal of all rows
wouldn't violate the constraint.

This therefore adds separate `delete_all` and `truncate` operations, so
we can apply the latter to the tables where the performance is poor (and
foreign keys permit).