drop_and_create_index requires only index name
The first implementation of this method required us to re-specify the
configuration of the index being dropped and recreated. This duplication
with the schema makes the introduction of errors possible, requiring
extra validation checks.

However, we can fetch the index config from ActiveRecord given only the
table and index names, so this change does exactly that instead.