Update arel from 8.0.0 to 9.0.0
This is required before being able to release a new version, as running
the `bundle exec rake release` command will fail:

```
$ bundle exec rake release
db_sanitiser 0.1.1 built to pkg/db_sanitiser-0.1.1.gem.
rake aborted!
There are files that need to be committed first.
```

The release command also runs `bundle install` which updates Gemfile.lock
and blocks a new release.

So let's commit the changes to unblock ourselves.

Here we're updating arel from 8.0.0 to 9.0.0 which matches what we run in the main futurelearn app. [1]

This is a dependency from the previous commit where we upgraded activerecord to 5.2.6.

[1]: https://github.com/futurelearn/futurelearn/blob/4f9da62c0526ccd6126454e5259819afa4f53729/Gemfile.lock#L142