Bump activerecord to 6.1.4
We've previously added some custom code to make this gem work with both
Rails 5.2 and Rails 6.1. [1]

We're now running on Rails 6.1.4 🎉 in the futurelearn app so we can
get rid of this code.

Before we can do that, we'll have to update this gem to use activerecord
6.1.4.

Notice this ends up getting rid of the `arel` gem. That's because this
is now bundled with activerecord. [2]

It also causes multiple depedencies to be updated:
- activemodel
- activerecord
- i18n
- tzinfo

And finally we're also adding zeitwerk as a dependency.

[1]: https://github.com/futurelearn/db_sanitiser/pull/6
[2]: https://github.com/rails/arel